requirements.txt
arrow==1.2.3
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
cookiecutter==1.7.3
Django==4.2.4
django-autoslug==1.9.9
djangorestframework==3.14.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.3
Pillow==10.0.0
poyo==0.5.0
psycopg==3.1.10
Pygments==2.16.1
python-dateutil==2.8.2
python-fire==0.1.0
python-http-client==3.3.7
python-slugify==8.0.1
pytz==2023.3
requests==2.31.0
rich==10.16.2
sendgrid==6.10.0
six==1.16.0
slug==2.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==2.0.4